requirements.txt
pytest-playwright == 0.4.4
allure-pytest ==  2.13.5
pytest == 8.1.1
playwright == 1.42.0
loguru == 0.6.0